requirements.txt
langchain
openai
beautifulsoup4
requests
python-dotenv
langchain-community